+++ README.md
@@ -171,6 +171,18 @@ docker-compose -f docker-compose.yml up
 | `spark_model` | 指定访问的模型版本       | lite指向Lite版本;4.0Ultra指向4.0 Ultra版本;                  |
 | `spark_token` | 星火大模型的Token        | 字符串密钥                                                   |
 
+#### 2.Grok
+
+##### 现在免费领取25刀的credits
+
+获取途径：https://console.x.ai/
+
+| 配置项       | 解释               | 备注                                         |
+| ------------ | ------------------ | -------------------------------------------- |
+| `grok_api`   | Grok服务的 BaseAPI | 默认为`https://api.x.ai/v1/chat/completions` |
+| `grok_model` | 指定访问的模型版本 | 默认为grok-beta                              |
+| `grok_token` | Grok的Token        | 字符串密钥                                   |
+
 ### ⚙️ GitHub Webhook 配置
 
 | 配置项 | 解释 | 备注 |
